configure with sql server

+++ database/migrations/2025_12_20_190533_remove_role_id_from_co_users.php
@@ -13,7 +13,9 @@ return new class extends Migration
     {
         Schema::table('co_users', function (Blueprint $table) {
             //
-            $table->dropColumn('role_id');
+            if (Schema::hasColumn('co_users', 'role_id')) {
+                $table->dropColumn('role_id');
+            }
         });
     }
